moved the Pratice3 component from components to templates folder
