+++ src/templates/Practice3.tsx
@@ -1,4 +1,4 @@
-import Input from "./Input";
+import Input from "../components/Input";
 
 export default function Practice2()
 {
